build-sys: honour the --enable/disable-dbus command line option.
All components that depend on D-Bus are now disabled unless the
--enable-dbus option (and --enable-gpl that the former depends
on) is specified on the command line. Moreover, the client D-Bus
interace has been turned into a plugin, so the core does not depend
or know anything about D-Bus any more.

Note that after converting to a plugin the D-Bus client interface
has not been tested at all yet...