daemon: support running with pa_glib_mainloop/GMainLoop.
To allow *temporarily* plugins to use gdbus, we need to be able
to run with GMainLoop and consequently to use pa_glib_mainloop.
This is now possible by setting the configuration variable
'gmainloop' to true either in the config file or by passing
the -s gmainloop=true on the command line.