wrt-media-client: added WRT media player client plugin
This plugin, at least in its current primitive form, is only here
temporarily. It registers and relays media playback commands by
emitting a D-Bus signal to a corresponding WRT plugin that catches
the signal and fires JS events for it.

This should be replaced by properly relaying (and extending/changing
as needed) the client protocol to the WRT. Relaying can, but does
not need to use D-Bus. We can have a dedicated WRT-specific plugin
using any socket-based IPC and a private dedicated protocol, but
we also could have, for instance, websocket bindings for the client
protocol and use that from the WRT.

The key thing is that a proper publish/subscribe protocol should be
used, so that the client can get proper notifications about changes
in the state of voice recognition. Currently the WRT (media) client
is not a proper client and for instance it does not know if VR/its
voice focus is active at all.