voice: add voice/text-to-speech API and a festival-based voice plugin.
This patch takes a first shot at adding voice rendering/text-to-speech
infra to the daemon. The basic API currently support

  - registering voice rendering backends + a set of supported voices
  - unregistering backends
  - querying the set of available voices (for language, dialect, speaker
    gender, age if known, and verbose voice description)
  - requesting text-to-speech rendering using a given voice
  - cancelling an ongoing or queued text-to-speech request

Rendering requests can also provide timeout/queuing preferences as
well as a rendering progress notification callback that can be used
to synchronize voice output with a traditional UI.

Note that, resource/policy integration for TTS is currently missing
altogether.

Also note that upstream libfestival (or more precisely the LISP
interpreter implementation of the Edinburgh speech tools libary) has
a bug which causes an immediate crash of any binary that is dynamically
linked against both libfestival and glibc. The actual bug is the leak
of the symbol 'backtrace' from slib.cc which collides with the backtrace
function of glibc. This symbol should have been declared static in slib
but was not. Because of the name collision, the dynamic linker resolves
it to the address of the function in glibc. In slib backtrace is an
ordinary read-write variable. Obviously bad things happen when slib tries
to change the value of that variable which now actually points to the
code segment of glibc.

The accompanying festival-loader plugin in this patch provides a kludge
to work around this problem without having to patch and recompile all
of festival/the speech tools library. Just load festival-loader before
attempting to load festival-voice and everything should work fine (with
fingers crossed).