daemon,native-client-api: first shot at socket-based activation.
This patch adds support for systemd socket-based activation to the
daemon/core and to the native client API plugin. The core part of
the patch adds a new command line option (--sockets/-S) that is
used to specify the order and list of configuration variables to
which the sockets passed in are assigned to. Multiple variables can
be specified as a comma-separated list the order of which must match
the order of Listen* directives in the assocaited systemd socket
file.

The native client API part changes to plugin to use any passed in
socket for the transport, or otherwise use the configured transport
address to create one.

This needs a corresponding fix to the murphy transports to work.