daemon: first shot at a basic common pulse interface.
In an attempt to provide a common PA interface and prevent
plugin-specific ones spreading like wildfire I lifted and
massaged espeak/pulse over as a common one to daemon. In
its current form it's not versatile enough to migrate all
the plugins over (eg. it provides nothing for recording or
tracking sinks) but it should be enough already for the
voice backend plugins.