sphinx: better logging and verbosity handling
Allow sphinx verbosity and logfile to be controlled by the
sphinx.verbose boolean and sphinx.log string keys. Also if
sphinx.log is set to 'srs', turn verbosity on and log sphinx
messages as murphy debug messages. sphinx warnings are turned
into murphy warnings and the rest (errors, critical errors
and fatal errors) are turned into murphy error messages.

Because we only get the originating file name and the line
number but not the function name where the message was
generated we have to cheat a bit and use 'sphinx' as the file
name and the file name as the originating function name. As a
consequence, you can see all sphinx INFO messages as murphy
debug messages by using the -d @sphinx command line option if
you have set sphinx.log to 'srs' in the config file (or have
passed -s 'sphinx.log=srs' to the daemon on the command line).