config: allow grouping config settings by key prefix.
You can now group configuration settings together with an
object-like notation. For instance, the following snippet
of configuration

    foo.blah = foobar
    foo.xyz  = zy
    foo.bar.foobar = xyzzy
    foo.bar.barfoo = yaddayadda

can be replaced with the equivalent snippet

    foo = {
        blah = foobar
        xyz  = zy
        bar = {
            foobar = xyzzy
            barfoo = yaddayadda
        }
    }

Internally both will be parsed to an identical representation.
The configuration parser is still the original primitive one.
Grouping has been bolted on top of it in a similarly primitive
manner as syntactic sugar. As a side-effect, if you make any
syntactic errors when using the grouping syntax (IOW nesting or
balancing errors), the resulting error messages probably will
not be very intuitive and in some cases might not help too much
in locating the actual error. Apologies...